watttime-emissions-index-scraper/requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.1.1
functions-framework==3.0.0
google-api-core==2.7.1
google-auth==2.6.2
google-cloud==0.34.0
google-cloud-pubsub==2.11.0
google-cloud-secret-manager==2.10.0
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.45.0
grpcio-status==1.45.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.7
requests==2.27.1
rsa==4.8
six==1.16.0
toml==0.10.2
urllib3==1.26.9
watchdog==1.0.2
Werkzeug==2.1.1
zipp==3.8.0
